Avoid all sorts of race conditions by waiting for stopped/started audio
Previously, the Nintendo DS could send a packet to the Supercard
indicating that it had consumed some samples of audio data while
the Supercard considered the audio to be stopped, and it handled
this by reducing the number of samples consumed.

Additionally, a data race could happen if:

* a new audio stream was started on the Supercard;
* the application submitted samples to the Supercard's audio queue,
  which were sent to the Nintendo DS;
* the Nintendo DS considered the audio to be on the previous stream;
* the Nintendo DS got the request to stop audio output and recreated
  its ring buffer.

In that case, the Supercard would have waited for the Nintendo DS
to give it a notification that the samples it sent were consumed,
which it would never have received (since the Nintendo DS now had
an empty buffer), leading to it permanently losing track of those
samples.

The fix is requested with a previously reserved byte in the
card_reply_hello structure, now called extensions.audio_status. If
the new MIPS side is glued to an ARM side that does not recognise
this byte, it will display the message

  Supercard protocol extension #0 not supported

and exit.