mips: Remove DMA code for communication with the Nintendo DS
DMA has potential to be faster if 32-byte DMA can be used, but as it
stands, because the address of video or audio data to be sent may be
aligned only to 4 bytes, 32-byte DMA cannot be used. 4-byte DMA is 4
times slower than 32-byte DMA; it makes more sense to send to the DS
with regular writes from the processor (and its data cache).