mips: Shorten the code for qsort; consistency fixes in stdlib.h
The helper procedure shellsort is modified to use simpler conditions,
both for tests and for loops, and reads more like insertion sort with
a stride instead of the abomination it once was.

It's also modified to use moves instead of swaps, holding one element
in memory and only writing that element at the correct place, and the
temporary element is stack-allocated instead of heap-allocated.

qsort's parameter names are shortened, bsearch's are modified to fit,
and both prototypes are modified in <stdlib.h>.