mips: Fix an oversight in the code that deals with PLL output
In the code that dealt with the instability of memory divider /1 by
doubling the PLL and setting the dividers to /2, formerly, only the
'hz' variable was updated, leaving the PLL multiplier in 'pll_m' as
before. Since only the 'pll_m' value meant anything, this meant the
effective frequencies were halved.