mips: Prepare the next video packet while the DS is decoding one
Previously, the Supercard only prepared a packet when the Nintendo DS
queried its send queue, and sent it when it was fully done. That made
it so the MIPS and ARM chips would never be working at the same time:

  MIPS  prep send|          |prep send|          |prep send ...
  ARM            |read query|         |read query|...

The situation changes with this commit to:

              v-- preparation and reading done in parallel
  MIPS  send|prep      |send|prep      |send|prep      |send ...
  ARM       |read query|    |read query|    |read query|...

This will allow some more processing-intensive video encodings to run
while the Nintendo DS is busy, to hopefully be done when the Nintendo
DS submits its next query.