mips: Do not require that applications have no small data in scds2_rules
The communication library and zlib must be built without data access
optimisations via $gp, so that they may work with applications which
cannot hold an address in $gp as well as applications that can. Only
applications which may hold an address in $gp may use optimisations;
the libraries will use full address loads.

However, the libraries are compiled with -G 256, which means that an
application may access some variables (e.g. stdout, stderr, or stdin
[currently unimplemented]) from them via $gp if it's built with a -G
value <= 256.