Add video encoding 1, which uses 8-bit palettes to transfer images
Because it does take some effort to compute the palette and determine
whether an image has under 252 unique colors, DS2_UseVideoCompression
must be called with a value of true for this encoding to be used.

When making the palette, a 32768-entry bit filter boosts the speed of
the search, making it O(1). The filter is then iterated over, to make
the final palette and a special reverse palette that maps 16-bit data
to 8-bit palette entries, allowing for palette searches in O(1).

Previous attempts used binary search and insertion sort. They weren't
fast enough.