mips: Add readdir_stat as an extension for efficiency in <ds2/ioext.h>
Applications are likely to take each entry returned by readdir() and
call stat() on it to know if the entry is a file or a directory. The
call to stat(), based purely on the file's name, will not know about
the ongoing directory iteration and will need to iterate through the
directory afresh to find the correct entry, resulting in O(n^2) time
to perform stat() in combination with readdir().

libfat's _FAT_dirnext function may fill a 'struct stat' directly, so
readdir_stat forwards the request from the application to libfat.