mips: Overhaul memmove
memmove is no longer just byte copy loops or a jump to memcpy if the
source and destination don't overlap.

It now handles word copies, aligned or not (via ULW and LW), as well
as (aligned) halfword copies. Cache line copies with PrepareForStore
prefetches are not done because they would unnecessarily bloat code:
four versions of the code would be required (forward and backward on
top of aligned and unaligned), and they would all need to check if a
cache line at the destination overlaps any data at the source, which
would clear some of the source data before it's copied.