mips: Add code for executable filtering (native) and unfiltering (MIPS)
Executable filtering is a technique that separates the many kinds of
data used to make up opcodes to help a compressor work on them. RISC
executable code is especially suited to filtering because it is very
regular:

* the filtering code can use tables to decode fixed-length opcodes;

* the unfiltering code has fewer possibilities for operand encodings
  to deal with, allowing for very small and efficient code.

The filter stops at the first word that doesn't look like an opcode,
then encodes the rest as the data segment.

In many cases, the filter actually enlarges the executable, but that
only matters if the filtered executable would fill over half the RAM
in the system once it's decompressed and before it's unfiltered.

The unfiltering code does not work when compiled at -Os; it has been
compiled at -O2.