mips: unfilter: Use more local variables
The loop that ended at 'instruction_count' re-read the value of the
global variable each time. It is rewritten to start at that value.

The opcode stream is used only in main. Set a local variable to the
stream pointer instead, allowing the compiler to use a callee-saved
register to hold its value.

This was suggested by Maarten ter Huurne.