mips: unfilter: Adjust the cache flushing behavior
Using Index (Writeback) Invalidate operations on the instruction cache
is undefined due to it ending up affecting the cache line that has the
instruction being run.

Now, only Hit Invalidate is used on the instruction cache, and only on
cache lines that are not being executed.

This issue was what caused the unfilterer not to work when compiled at
-Os; since it was fixed, the Makefile is edited to compile at -Os.