mips: unpacker: Adjust the cache flushing behavior
Using Index (Writeback) Invalidate operations on the instruction cache
is undefined due to it ending up affecting the cache line that has the
instruction being run.

Now, only Hit Invalidate is used on the instruction cache, and only on
cache lines that are not being executed.

Unneeded Coprocessor 0 operations are also deleted.