mips: Allow interrupt nesting for unhandled exception reports
As written in the assembly code prior to this change, the code that
sends exception reports to the Nintendo DS must run with interrupts
enabled. However, if an interrupt was triggered while copying saved
registers from the stack (at 0x8000_1F40) and it reset the stack to
0x8000_1FE0 to process the interrupt, _unhandled_exception's values
were corrupted.

The symptom was an incorrect exception report if any data was still
in the queue to be sent to the Nintendo DS and this was more likely
as the CPU was downclocked.

This change allows one level of interrupt nesting by checking if sp
is already at 0x8000_1FE0 or below and using it verbatim if so. The
exception handlers, by contrast, do not check whether sp is already
in range.